api/requirements.txt
fastapi
uvicorn[standard]
python-keycloak
python-dotenv
pydantic
python-jose[cryptography]
requests
